Update startup.sh
minor update

+++ startup.sh
@@ -4,8 +4,7 @@ ip route change default via 172.17.42.254
 
 ulimit -n 2048
 
-echo cp -Ra /seed/${CONTAINER_TYPE}/ /server/
-cp -Ra /seed/${CONTAINER_TYPE}/ /server/
+if [ "$( find /server/ -type f | wc -l )" -lt "1" ]; then cp -Ra /seed/${CONTAINER_TYPE}/ /server/; fi
 /root/steamcmd/steamcmd.sh +login anonymous +force_install_dir /server +app_update 294420 +quit
 
 sed -i "s%^  <property name=\"ServerPort\"[ \t]*value=\"[0-9]*\"/>%  <property name=\"ServerPort\" value=\"${PORT_26900}\"/>%" /server/serverconfig.xml
